refactor: 'kpm registry login/logout' -> 'kpm login/logout'.
