+++ test/e2e/test_suites/kpm/no_args/login_reg_without_args/test_suite.input
@@ -1,1 +1,1 @@
-kpm registry login ghcr.io
+kpm login ghcr.io
