+++ test/e2e/test_suites/kpm/no_args/login_reg_without_password/test_suite.input
@@ -1,1 +1,1 @@
-kpm registry login -u aaaa ghcr.io
+kpm login -u aaaa ghcr.io
