+++ test/e2e/test_suites/kpm/no_args/login_reg_without_username/test_suite.input
@@ -1,1 +1,1 @@
-kpm registry login -p aaaa ghcr.io
+kpm login -p aaaa ghcr.io
