+++ test/e2e/test_suites/kpm/no_args/login_with_no_args/test_suite.input
@@ -1,1 +1,1 @@
-kpm registry login
+kpm login
