+++ test/e2e/test_suites/kpm/no_args/logout_reg_with_invalid_reg/test_suite.input
@@ -1,1 +1,1 @@
-kpm registry logout invalid_registry
+kpm logout invalid_registry
