+++ test/e2e/test_suites/kpm/no_args/logout_reg_without_args/test_suite.input
@@ -1,1 +1,1 @@
-kpm registry logout
+kpm logout
